refactor(sentence): moved parsePlaceholderNotation to utils
